04 auth

+++ controllers/contacts.js
@@ -1,14 +1,22 @@
 import { HttpError } from '../helpers/HttpError.js';
-import { Contact, addSchema, patchSchema, putSchema } from '../models/contact.js';
+import { Contact } from '../models/contact.js';
 
 export const getAll = async (req, res) => {
-    const result = await Contact.find();
+    const { _id: owner } = req.user;
+    const { page = 1, limit = 20, favorite = null } = req.query;
+    const skip = (page - 1) * limit;
+    const condition = { owner };
+    if (favorite !== null) {
+        condition.favorite = favorite;
+    }
+    const result = await Contact.find(condition, '-createdAt -updatedAt', { skip, limit }).populate('owner', 'email subscription');
     res.status(200).json(result);
 };
 
 export const getById = async (req, res) => {
+    const { _id: owner } = req.user;
     const { contactId } = req.params
-    const result = await Contact.findOne({ _id: contactId });
+    const result = await Contact.findOne({ owner, _id: contactId });
     if (!result) {
         throw HttpError(404, "Not found")
     }
@@ -16,17 +24,15 @@ export const getById = async (req, res) => {
 };
 
 export const postContact = async (req, res) => {
-    const { error } = addSchema.validate(req.body)
-    if (error) {
-        throw HttpError(400, error.message)
-    }
-    const result = await Contact.create(req.body);
+    const { _id: owner } = req.user;
+    const result = await Contact.create({...req.body, owner});
     res.status(201).json(result);
 };
 
 export const deleteContact = async (req, res) => {
-    const { contactId } = req.params;
-    const result = await Contact.findByIdAndDelete(contactId);
+    const { _id: owner } = req.user;
+    const { contactId } = req.params
+    const result = await Contact.findByIdAndDelete({ owner, _id: contactId });
     if (!result) {
         throw HttpError(404, "Not found")
     }
@@ -36,12 +42,9 @@ export const deleteContact = async (req, res) => {
 };
 
 export const putContact = async (req, res) => {
-    const { error } = putSchema.validate(req.body)
-    if (error) {
-        throw HttpError(400, "Missing fields")
-    }
+    const { _id: owner } = req.user;
     const { contactId } = req.params;
-    const result = await Contact.findByIdAndUpdate({ _id: contactId }, req.body, { new: true });
+    const result = await Contact.findOneAndUpdate({ owner, _id: contactId }, req.body, { new: true });
     if (!result) {
         throw HttpError(404, "Not found")
     }
@@ -49,15 +52,11 @@ export const putContact = async (req, res) => {
 };
 
 export const updateStatusContact = async (req, res) => {
-    const { error } = patchSchema.validate(req.body)
-    if (error) {
-        throw HttpError(400, error.message)
-    }
+    const { _id: owner } = req.user;
     const { contactId } = req.params;
-    const { body } = req;
-    const result = await Contact.findByIdAndUpdate(contactId, body, { new: true });
+    const result = await Contact.findOneAndUpdate({ owner, _id: contactId }, req.body, { new: true });
     if (!result) {
         throw HttpError(404, "Not found")
     }
     res.status(200).json(result);
-};
+}
